Set up styles for nav-bar
Made the main activity get rid of the original nav-bar and make it use the one created. Also created a string file for SDK 21, hopeflly it can solve the issue with the widget needing SDK 21. For some reason the default style.xml file was not generated, so I created one. It is giving errors though.